# openapi.yaml
openapi: 3.0.3
info:
  title: NexusID API
  version: 0.1.0
  description: OAuth2/OIDC, admin APIs, and user portal APIs.
servers:
  - url: http://127.0.0.1:4000
paths:
...
  /api/admin/access-requests:
    get:
      summary: List access governance requests
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected, expired, cancelled]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: Access request list
    post:
      summary: Create access governance request
      responses:
        '201':
          description: Access request created
  /api/admin/access-requests/{id}/approve:
    post:
      summary: Approve access governance request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Access request approved
  /api/admin/access-requests/{id}/reject:
    post:
      summary: Reject access governance request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Access request rejected
  /api/admin/access-requests/process-expirations:
    post:
      summary: Process expired approved access requests and revoke granted assignments
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                dryRun:
                  type: boolean
      responses:
        '200':
          description: Expiration processing summary
  /api/admin/access-reviews/campaigns:
    post:
      summary: Create an access review campaign and generate review items from current assignments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                dueAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Access review campaign created
  /api/admin/access-reviews/campaigns/{id}:
    get:
      summary: Get access review campaign with generated items
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Access review campaign details
  /api/admin/access-reviews/items/{id}/decision:
    post:
      summary: Record a decision for an access review item and emit attestation evidence metadata
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [certified, revoked]
                rationale:
                  type: string
      responses:
        '200':
          description: Access review item updated
...
  /api/admin/access-requests/stalled:
    get:
      summary: List stalled pending access requests that have exceeded the SLA threshold
      parameters:
        - in: query
          name: stalledAfterMinutes
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: List of stalled access requests
...
  /api/admin/connectors:
    get:
      summary: List connectors
      tags: [Connectors]
      responses:
        '200':
          description: Connector list
    post:
      summary: Create connector
      tags: [Connectors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type, config]
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [ldap, scim, csv, sql, custom]
                config:
                  type: object
                schedule:
                  type: string
      responses:
        '201':
          description: Connector created
  /api/admin/connectors/{id}:
    get:
      summary: Get connector
      tags: [Connectors]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Connector detail
        '404':
          description: Connector not found
    patch:
      summary: Update connector
      tags: [Connectors]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [ldap, scim, csv, sql, custom]
                status:
                  type: string
                  enum: [active, inactive, error]
                config:
                  type: object
                schedule:
                  type: string
      responses:
        '200':
          description: Updated connector
        '404':
          description: Connector not found
    delete:
      summary: Delete connector
      tags: [Connectors]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Connector deleted
  /api/admin/connectors/{id}/sync:
    post:
      summary: Trigger connector sync
      tags: [Connectors]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Connector sync run accepted
        '400':
          description: Connector inactive
        '404':
          description: Connector not found
  /api/admin/connectors/{id}/runs:
    get:
      summary: List connector runs
      tags: [Connectors]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Connector run history
        '404':
          description: Connector not found
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT